requirements-dev.txt
mypy
types-requests
PyQt5-stubs
